requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
openai==2.17.0
python-docx==1.1.2
python-pptx==0.6.23
PyMuPDF==1.24.14
psycopg[binary]==3.2.9
